Add dev task to facilitate my workflow
For reason that I don't quite understand, when I run the `rails 
db:reset` command that suppose to drop, create and populate the database 
with the seeds, I was getting the message telling me to run the migrate 
first. I run it and keep getting an error that my table on database 
already exists. So, my solution was creating this task that actually do 
the same, but in separated steps.