Update to Rails 7 o/
Firstly, I didn't notice that I started the project not using the latest 
Rails version. Also, the reason that I took me to update the Rails was 
specifically because I saw the new Active Record validator called 
`comparison`. I found that it could fit in one of mine `Appointment` 
validations so that's why I updated the Rails.